+++ R/tool-r-help.R
@@ -96,10 +96,3 @@ tool_r_help <- function(allowed_packages = c("base", "stats", "utils",
   )
 }
 
-#' @rdname tool_r_help
-#' @param ... Arguments passed to [tool_r_help()].
-#' @export
-r_help_tool <- function(...) {
-  lifecycle::deprecate_warn("0.3.0", "r_help_tool()", "tool_r_help()")
-  tool_r_help(...)
-}